backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
requests==2.28.2
python-multipart==0.0.5
Werkzeug==2.0.3 